requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
mysql-connector-python==9.0.0
requests==2.32.3
pdfplumber==0.11.7
pandas==2.3.2
Pillow==11.3.0
pypdfium2==4.30.0
